webapp/requirements.txt
annotated-types==0.5.0
anyio==4.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.0
h11==0.14.0
idna==3.4
install==1.3.5
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2